pyproject.toml
[project]
name = "finclarify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "gtts>=2.5.4",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.6",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
    "streamlit-lottie>=0.0.5",
    "uvicorn>=0.40.0",
]
